requirements.txt
discord.py>=2.3.0
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
python-dotenv>=1.0.0